requirements.txt
customtkinter>=5.2.0
tkinterdnd2>=0.4.0
pillow>=10.0.0
reportlab
matplotlib
pymupdf>=1.24.0
